New Theme System
Backwards compatible with existing styles

Uses ElementName, ID, and/or className. For backwards compatibility, className is set using the Style attribute.

Multiple style classes can be used on a view using space delimiting: <MyView Style="Class1 Class2 Class3">

Xuml Theme can now use <Style> element, which defines a style that is not constraind by an element name (view type).

Attributes in styles that are not present in a view are ignored.

BasedOn attribute is still available for backwards compatibility, but not necessary given the ability to combine multiple styles.