Parent/Child Style Relationships
Add ability to specify parent/child relationships in styles. For example, If you create a Button style, you can add a child Label style. The child Label style will only be applied if the Label is a child of a Button element that matches the parent Button style selector.

Added additional comments.