ObservableList Improvements
Implemented more perfect MVC pattern. All data model manipulation should be done through ObservableList or the ObservableItem's that wrap the list values. Views should primarily be responsible for rendering the model.

ObservableList is a list of ObservableItem's. For convenience, it acts as a list of whatever value type specified. ObservableItems can be passed to a views Item field.

Item selection flag is now stored by the ObservableItem as part of the data model. Changing the selection can be done through the IsSelected property of the ObservableItem. It can also be achieved through the SelectedItem and SelectedIndex properties of both the ObservableList and List view.

 The Item field is now an ItemDataViewField type, which acts as an _object view field type. However, if an IObservableItem implementation is passed into it, it will set the IObservableItem's Value property as its own value and store the IObservableItem seperately in the view fields ObservableItem property.

 The IObservableItem has events that the ListItem view can subscribe to. These events allow the ListItem to be notified directly of changes to its assigned data model (IObservableItem). The ListItem view now handles some of the responsibilities that the List view formerly handled.

 Changes to the IObservableItem are still passed as events to the List view, which handles presentation changes accordingly.