Fix: Some duplicate styles not applied
Theme.cs was incorrectly selecting root styles based on not having a parent instead of based on if it has properties or not.

CSSThemeLoader.cs ensure that duplicate styles are merged.

CssSelector.css ensure that Raw property is exactly the same across multiple instances of matching selectors regardless of minor detail differences such as additional spaces in combinators.