Binding Improvements
Added ^ binding modifier. This modifier will look for the first layout parent view that has the specified binding field all the way to the layout root view.

The # binding modifier has been altered. It will look for the specified binding field on the local element as before. However, if the local element does not have the specified field, it will begin searching for the first layout parent that has the field. It will not search farther than the local elements non-layout parent.

If the ^ and # modifier are combined, the behavior will be that of the # modifier, except it will search all the way to the layout root instead of stopping at the non-layout parent of the local element.

Added exception handlers when notifying observers in order to prevent one error from effecting other observers from being notified. It should also help narrow the cause of uncaught exceptions.

ParseBindingString method improved to reduce potential number of string created.

General code cleanup and refactoring.