Add separate layout calculator classes for views.
Seperate layout calculators allow layout code to be easily.

LayoutCalculator is an abstract implementation that all layout calculators must extend.

DefaultLayoutCalculator does nothing. It is the default layout calculator that is used by View if the LayoutCalculator property is not set.

UniformStackLayoutCalculator is functionality taken from the StackPanel view.

GroupLayoutCalculator is from the Group and List view. Improved so that child items are not forced to be, for example, TopLeft aligned when vertical. Child View's can be Left, Center, or Right aligned, even when wrapped. However, the alignment set on the child View's are no longer considered.

Moved LayoutChangeContext, LayoutData, and LayoutRectCalculator to Layout folder.

Updated Views that override CalculatorLayoutChanges method.

Updated List to ensure layout update when resizing view port in Unity editor.

Cleanup code in touched areas.