Improved Layout Performance
Layout Retries not needed; Removed.

(LayoutChangeContext) Add addition calculate methods to LayoutChangeContext: CalculateAsParent, CalculateAsChild

(LayoutChangeContext) The view that starts layout calculation is the source view and uses the Calculate method. Its parent and ancestors are calculated as parent via the CalculateAsParent method. Its children and descendants are calculated as child via the CalculateAsChild method. Certain aspects of calculation are performed depending on method used in order to prevent doing more than is required.

(View) _isLayoutChanged value to false when calculating as child or parent to ensure view is not calculated again as source.