Improve dependent value observer notifications.
List view calls parent views Field.NotifyDependentValueObservers method when the data model is modfied or selection is changed. This ensures bindings that target the List are notified of changes automatically.

Added ObservableListViewField class. Listens for modifications or changes and notifies the owner view dependent value observers. _IObservableList and GenericObservableList extend ObservableListViewField.

In ViewFields.NotifyDependentValueObservers method: Ensure notification of value observers are performed for mapped fields so long as the next object in the field path is a view. This fixes a problem where bindings do not see changes when they are bound to a field and the notification is performed on a field in a parent view that is mapped to the bounded field. In such a case, the field was not notified because it was mapped so the binding was never notified of changes.

ViewFieldPathInfo updated to ensure the target view returns the target view (if any) when mapped. Renamed ViewType field to SourceViewType field.

Fixed ViewFieldPathInfo.GetValue method so that the last field in the path is always considered to have a value (null is considered a valid value). This fixes issue where bindings are not updated if the binding source value is set to null.

Refactored ViewField.ParentView to ViewField.OwnerView. This is to make a distinction between a view that is a parent of another view and a view that owns an instance of a non-view object so that when the two are used in the same scope, there will be no naming conflict or misunderstanding.