Add DefaultBackgroundImage field to UIView
Add DefaultSprite field to Image view.

These are to make it easy to set a global (or context specific) default image to show (via stylesheets) when the BackgroundImage value is set to null. Usefull when setting image via binding to a data source and you need a default image to show when the binding value is null.