Add ViewFieldValue struct, Observable notification handler in View
ViewFieldValue is used to hold context for a single attempt at settings a fields value. This is to reduce the number of method arguments in ViewFields.SetValue and ViewFieldData.SetValue. It also makes it easier to read and understand what settings are being used when looking at a SetValue call.

Added ValueObserversNotified callback method to View. This is called whenever a view fields value observers are notified of a value change.