Add AspectRatio view field.
AspectRatio view field sets the desired aspect ratio of a view. The aspect ratio is maintained on the view whenever the size of the view is changed. The width or height will be changed depending on which value can be changed. Only percent values can be changed for aspect ratio purposes. If a size is set in pixels, it cannot be adjusted. Therefore, if both width and height are set in pixels, the AspectRatio value will have no effect.

The AspectRatio value can be written in Xaml or CSS stylesheet using a single ratio value (i.e. 0.75) where the ratio value is the width divided by height. It can also be written as two values that represent the X and Y aspect ratio (like with TV's and monitors). The two values can be seperated with a space, a comma, or a colon (i.e. 16:9).

The AspectRatio is ignored unless the value is set.

Added ElementAspectRatio data object.

Added ElementAspectRatioConverter.