Allow view to be moved by margin if there is room.
Previously the view is the specified size minus margins. This can cause the view to be smaller than desired or even collapse completely if the a margin is larger than the size of the view.

LayoutRectCalculator changed to allow margin to displace view if there is room. If the size of the view is specified in pixels, the view size will be maintained against the margin. The margin side that matches the alignment will be maintained while the opposite margin will not be honored. For percent based margins, the size of the view may be reduced in order to honor the margins.