Fix infinite loop on ViewField notifications.
Fixed infinite loop in List where, under certain conditions, selecting an item can cause the ItemsChanged handler to be called which then causes the ItemChanged handler to be queued again. This is caused by the List notifying the parent that the List in general is updated. If the parent has a field mapped to the lists Items field, the field will be notified again.

Solution is to allow turning of notification of mapped view fields via an additional argument on ViewFields.NotifyDependentValueObservers called "includeMapped". This should be set to false whenever calling the parent views ViewFields.NotifyDependentValueObservers to prevent the notification from coming back to the source view.

List view is the view that uses the method ViewFields.NotifyDependentValueObservers in such a manner in order to aid with bindings to selected items. It has been updated to fix the mentioned issue.