Layout improvements.
Improvements in layout of List, DataGrid, and layout in general.

Layout updates via LayoutChangeContext are no longer propagated. The context itself is not instantiated within the view. A single context is responsible for all view layout updates and is instantiated in ViewPresenter. Any view marked dirty will be updated during the ViewPresenters late update in reverse breadth order.

Fixed margin handling in LayoutCalculator and LayoutRectCalculator to be simpler. The pixel size (width/height) of the view is still honored even if margins are specified. The effective displacement of a view is its size plus margins. When using percent sizes, the size may be changed in order to honor the margins.

Removed "Spacing", "HorizonalSpacing", and "VerticalSpacing" view fields from Group, List, DataGrid. Better spacing control can be achieved by setting the Margin field on the individual items. GroupLayoutCalculator and DataGridRowLoayoutCalculator updated to make this possible.